Update all dependencies, improve docker build
The docker build runner scripts were losing the quotes on arguments
leading to it being impossible to process videos with spaces in their
filenames.

Additionally, remove the max version from numpy in requirements.txt. The
only reason it was there is because the released version of sk-video is
years out of date, and does not support modern numpy. The sk-video
update required a change in how it is imported in inference_video.py.